requirements.txt
google-genai
google-auth
rich
